+++ react/FilterNavigator.js
@@ -33,6 +33,7 @@ const FilterNavigator = ({
   brands = [],
   loading = false,
   filters = [],
+  preventRouteChange = false,
   hiddenFacets = {},
 }) => {
   const { isMobile } = useDevice()
@@ -77,8 +78,15 @@ const FilterNavigator = ({
 
   if (isMobile) {
     return (
-      <div className={filterClasses}>
-        <FilterSidebar filters={filters} tree={tree} priceRange={priceRange} />
+      <div className={styles.filters}>
+        <div className={filterClasses}>
+          <FilterSidebar
+            filters={filters}
+            tree={tree}
+            priceRange={priceRange}
+            preventRouteChange={preventRouteChange}
+          />
+        </div>
       </div>
     )
   }
@@ -95,14 +103,22 @@ const FilterNavigator = ({
             <FormattedMessage id="store/search-result.filter-button.title" />
           </h5>
         </div>
-        <SelectedFilters filters={selectedFilters} />
+        <SelectedFilters
+          filters={selectedFilters}
+          preventRouteChange={preventRouteChange}
+        />
         <DepartmentFilters
           title={CATEGORIES_TITLE}
           tree={tree}
           isVisible={!hiddenFacets.categories}
           onCategorySelect={navigateToFacet}
+          preventRouteChange={preventRouteChange}
         />
-        <AvailableFilters filters={filters} priceRange={priceRange} />
+        <AvailableFilters
+          filters={filters}
+          priceRange={priceRange}
+          preventRouteChange={preventRouteChange}
+        />
       </div>
       <ExtensionPoint id="shop-review-summary" />
     </Fragment>